split training tests into separate file
